feat: add image download btn component for moaBox and letter, add id name for capture area
